+++ src/app/templates/GeneralSettings.tsx
@@ -68,7 +68,7 @@ const GeneralSettings: FC = () => {
   return (
     <SubPageLayout data-testid="general-settings">
       <SubPageSection>
-        <ListGroup>
+        <ListGroup surface="outline">
           <ListRow
             title={t('theme')}
             trailing={
@@ -96,7 +96,7 @@ const GeneralSettings: FC = () => {
       </SubPageSection>
 
       <SubPageSection footnote={t('delegateProofSettingsDescription')}>
-        <ListGroup>
+        <ListGroup surface="outline">
           <SettingToggle
             checked={delegateEnabled}
             onChange={handleDelegateChange}
@@ -108,7 +108,7 @@ const GeneralSettings: FC = () => {
       </SubPageSection>
 
       <SubPageSection footnote={t('autoConsumeSettingsDescription')}>
-        <ListGroup>
+        <ListGroup surface="outline">
           <SettingToggle
             checked={consumeEnabled}
             onChange={handleAutoConsumeChange}