+++ src/screens/developer-settings/DeveloperSettings.tsx
@@ -305,7 +305,7 @@ const DeveloperSettings: React.FC<DeveloperSettingsProps> = ({ readOnly = false 
         />
       </SubPageSection>
 
-      <ListGroup>
+      <ListGroup surface="outline">
         <ListRow
           title={t('devAllowNoGuardian')}
           disabled={readOnly}